delete endpoint find user for form
